Revert "Contextual hints"
