+++ apps/i18n/common/en_us.json
@@ -201,7 +201,6 @@
   "print": "Print",
   "puzzleTitle": "Puzzle {puzzle_number} of {stage_total}",
   "readonlyWorkspaceHeader": "View only: ",
-  "recommendedBlockContextualHintTitle": "Try using a block like this to solve the puzzle.",
   "repeat": "repeat",
   "resetProgram": "Reset",
   "rotateText": "Rotate your device.",
