+++ apps/src/authoredHintUtils.js
@@ -1,6 +1,4 @@
 /* global marked */
-var parseXmlElement = require('./xml').parseElement;
-var msg = require('./locale');
 
 /**
  * @overview A helper class for all actions associated with the Authored
@@ -224,30 +222,9 @@ authoredHintUtils.submitHints = function (url) {
 };
 
 /**
- * Generates contextual hints as used by StudioApp from Blockly XML
- * @param {String[]} - array of Blockly Block XML strings
- * @return {AuthoredHint[]}
- */
-authoredHintUtils.createContextualHintsFromBlocks = function (blocks) {
-  var hints = blocks.map(function (block) {
-    var xmlBlock = parseXmlElement(block.blockDisplayXML);
-    var blockType = xmlBlock.firstChild.getAttribute("type");
-    return {
-      content: marked(msg.recommendedBlockContextualHintTitle()),
-      block: xmlBlock,
-      hintId: "recommended_block_" + blockType,
-      hintClass: 'recommended',
-      hintType: 'contextual',
-      alreadySeen: block.seen
-    };
-  });
-  return hints;
-};
-
-/**
  * Generates authored hints as used by StudioApp from levelbuilder JSON.
  * @param {string} - JSON representing an array of hints
- * @return {AuthoredHint[]}
+ * @return {Object[]}
  */
 authoredHintUtils.generateAuthoredHints = function (levelBuilderAuthoredHints) {
   if (!marked) {
