+++ apps/src/feedback.js
@@ -175,19 +175,7 @@ FeedbackUtils.prototype.displayFeedback = function(options, requiredBlocks,
 
   var onlyContinue = continueButton && !againButton && !previousLevelButton;
 
-  // get a list of all missing recommended blocks, and display them as
-  // contextual hints when the dialog closes. If the user views one of
-  // the blocks in the dialog, it is removed from this list.
-  var missingRecommendedBlockHints = this.getMissingBlocks_(recommendedBlocks, 1)
-    .blocksToDisplay
-    .map(function (block) {
-      block.seen = false;
-      return block;
-    });
-  var onHidden = onlyContinue ? options.onContinue : function () {
-    this.studioApp_.displayMissingBlockHints(missingRecommendedBlockHints);
-  }.bind(this);
-
+  var onHidden = onlyContinue ? options.onContinue : null;
   var icon;
   if (!options.hideIcon) {
     icon = canContinue ? this.studioApp_.winIcon : this.studioApp_.failureIcon;
@@ -245,9 +233,6 @@ FeedbackUtils.prototype.displayFeedback = function(options, requiredBlocks,
       // Remove "Show hint" button.  Making it invisible isn't enough,
       // because it will still take up space.
       hintRequestButton.parentNode.removeChild(hintRequestButton);
-
-      // mark the corresponding block hint as seen
-      missingRecommendedBlockHints[0].seen = true;
     } else {
 
       // Generate a generic feedback message to display when we show the
@@ -261,8 +246,6 @@ FeedbackUtils.prototype.displayFeedback = function(options, requiredBlocks,
 
       // If the user requests the hint...
       dom.addClickTouchEvent(hintRequestButton, function () {
-        // mark the corresponding block hint as seen
-        missingRecommendedBlockHints[0].seen = true;
 
         // Swap out the specific feedback message with a generic one.
         var parentNode = feedbackMessage.parentNode;
