+++ apps/src/templates/HintsDisplay.jsx
@@ -1,6 +1,3 @@
-var msg = require('../locale');
-var Hint = require('./Hint.jsx');
-
 /**
  * @overview React Component for displaying Authored Hints in the
  * Instructions dialog. Any hints the user has already requested to see
@@ -34,34 +31,6 @@ module.exports = React.createClass({
     });
   },
 
-  /**
-   * Many of our hints include Blockly blocks. Unfortunately, Blockly
-   * BlockSpaces have a real problem with being created before they are
-   * in the DOM, so we need to inject this BlockSpace outside of our
-   * React render method once we're confident that this component is in
-   * the DOM.
-   */
-  injectBlocklyHint: function (hint) {
-    var ref = this.refs[hint.hintId];
-    ref.injectBlocklyHint();
-  },
-
-  componentDidMount: function () {
-    // now that we're in the DOM, we can render our Blockly blocks for
-    // those hints that have them
-    this.props.seenHints.filter(function (hint) {
-      return hint.block;
-    }).forEach(this.injectBlocklyHint);
-  },
-
-  componentDidUpdate: function () {
-    // if our update has us showing a new hint, make sure to render the
-    // block if it has one
-    if (this.state.showNextUnseenHint && this.props.unseenHints[0].block) {
-      this.injectBlocklyHint(this.props.unseenHints[0]);
-    }
-  },
-
   render: function () {
     var hintsToShow = this.props.seenHints;
     if (this.state.showNextUnseenHint) {
@@ -74,7 +43,7 @@ module.exports = React.createClass({
           <h1>{ this.props.hintReviewTitle }</h1>,
           <ul>
             {hintsToShow.map(function (hint) {
-              return <Hint hint={hint} ref={hint.hintId} />;
+              return <li dangerouslySetInnerHTML={{ __html : hint.content }} style={{ marginBottom: '12px' }}/>;
             })}
           </ul>
       ];
@@ -83,9 +52,9 @@ module.exports = React.createClass({
     var viewHintButton;
     if (!this.state.showNextUnseenHint && this.props.unseenHints && this.props.unseenHints.length) {
       viewHintButton = (
-        <button id="hint-button" onClick={ this.viewHint } className="lightbulb-button">
+        <button id="hint-button" onClick={ this.viewHint }>
           <span dangerouslySetInnerHTML={{ __html: this.props.lightbulbSVG }} />
-          {msg.hintSelectNewHint()}
+          Get a new hint
         </button>
       );
     }
