+++ apps/src/templates/buttons.html.ejs
@@ -34,8 +34,7 @@
     </div>
   <% } else { -%>
     <% if (data.shouldPromptForHint) { -%>
-      <button id="hint-request-button" class="lightbulb-button">
-        <span><%- include lightbulb.svg.ejs %></span>
+      <button id="hint-request-button" class="launch">
         <%= msg.hintRequest() %>
       </button>
     <% } %>
