+++ apps/style/common.scss
@@ -29,21 +29,6 @@ $common_root: '/blockly/media/common_images/';
   transform: scale(-1, 1);
 }
 
-@keyframes wiggle {
-  0% { transform: rotate(0deg); }
-  25% { transform: rotate(-20deg); }
-  75% { transform: rotate(20deg); }
-  100% { transform: rotate(0deg); }
-}
-
-.animate-hint {
-  animation-name: wiggle;
-  animation-duration: 200ms;
-  transform-origin: 50% 50%;
-  animation-iteration-count: 3;
-}
-
-
 body {
   @include font;
   background-color: #F2F2F2;
