+++ dashboard/app/assets/stylesheets/application.css.scss
@@ -1948,7 +1948,7 @@ $modal-no-icon-gap: 22px;
   min-width: 100px;
 }
 
-.lightbulb-button {
+#hint-button {
   background-color: $lightest_gray;
   border: 1px solid $light_gray;
   color: black;
